removed AIRRSequenceImport, moved functionality
